Accept fractional `vram_gb` for local models
close #10

The Discover UI writes model size estimates in GiB with two
decimals, and the `u32` schema rejected every non-whole-GiB value.
Change the `LocalModelConfig` field `vram_gb` from `Option<u32>` to
`Option<f64>` and validate that a set value is finite and positive.

- The new type matches the `[[stt_model]]` estimate, so the
  co-residency total in `validate.rs` adds it without a `f64::from`
  conversion.
- Validation rejects a non-finite or non-positive value with the
  error `vram_gb must be finite and greater than zero`; a test
  covers `0.0`, `-1.0`, `nan`, and `inf`.
- Integer TOML input still parses: a test pins `vram_gb = 14` to
  `Some(14.0)`.
- The `vram_gb` field on the dominion table stays `u32`; only the
  local model estimate accepts fractions.

+++ crates/promptforge-gateway-config/src/config.rs
@@ -354,9 +354,9 @@ pub struct LocalModelConfig {
     #[serde(default = "default_parallel")]
     parallel: u32,
     /// VRAM footprint estimate in gibibytes for the dominion co-residency
-    /// check.
+    /// check. Fractional values are accepted, matching `[[stt_model]]`.
     #[serde(default)]
-    vram_gb: Option<u32>,
+    vram_gb: Option<f64>,
     /// Context window size in tokens (`--ctx-size`).
     context: u32,
     /// Whether thinking tokens are never, always, or switchably available.
